+++ sync-for-commerce/src/codat_sync_for_commerce/models/operations/get_sync_flow_url.py
@@ -8,8 +8,8 @@ from codat_sync_for_commerce.utils import (
     QueryParamMetadata,
 )
 import pydantic
-from typing import Optional, TypedDict
-from typing_extensions import Annotated, NotRequired
+from typing import Optional
+from typing_extensions import Annotated, NotRequired, TypedDict
 
 
 class GetSyncFlowURLRequestTypedDict(TypedDict):