+++ sync-for-commerce/src/codat_sync_for_commerce/models/operations/update_connection_authorization.py
@@ -8,8 +8,8 @@ from codat_sync_for_commerce.utils import (
     RequestMetadata,
 )
 import pydantic
-from typing import Dict, Optional, TypedDict
-from typing_extensions import Annotated, NotRequired
+from typing import Dict, Optional
+from typing_extensions import Annotated, NotRequired, TypedDict
 
 
 class UpdateConnectionAuthorizationRequestTypedDict(TypedDict):