+++ sync-for-commerce/src/codat_sync_for_commerce/models/shared/brandingimage.py
@@ -3,8 +3,8 @@
 from __future__ import annotations
 from .imagereference import ImageReference, ImageReferenceTypedDict
 from codat_sync_for_commerce.types import BaseModel
-from typing import Optional, TypedDict
-from typing_extensions import NotRequired
+from typing import Optional
+from typing_extensions import NotRequired, TypedDict
 
 
 class BrandingImageTypedDict(TypedDict):