+++ sync-for-commerce/src/codat_sync_for_commerce/models/shared/groupinglevels.py
@@ -8,8 +8,8 @@ from .invoicelinelevelselection import (
 )
 from codat_sync_for_commerce.types import BaseModel
 import pydantic
-from typing import Optional, TypedDict
-from typing_extensions import Annotated, NotRequired
+from typing import Optional
+from typing_extensions import Annotated, NotRequired, TypedDict
 
 
 class GroupingLevelsTypedDict(TypedDict):